manually uploading the files I changed into github (#42)
* manually uploading the files I changed into github

the new container refuses to let me authenticate so i just downloaded them and put them here. I only am uploading what I changed so that it wont interfere with anything else.

* Reversed previous merge commit, but included changes. Moved tests from bottom of views.py to tests.py

* Reversed previous merge commit, but included changes. Moved tests from bottom of views.py to tests.py

* Resolved formatting issue in data theme light

* Removed CalendarWeek reference in imports

---------
