+++ project/home/templates/base_template.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
-    <!-- Ensure that the static folder is available for use on this template -->
+    <!-- ensure that static folder is available for use on this template -->
 {% load static %}
 {% load bootstrap5 %}
 {% bootstrap_css %}
@@ -8,48 +8,46 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <!-- Use bootstrap for sticky navigation bar -->
+    <!-- Used bootstrap to make sure sticky navigation bar with game-logo on the left and routes to other pages -->
     <link rel="stylesheet" href="{% static 'styles-v2.css' %}">
     <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" rel="stylesheet">
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     
     <title>
-        {% block title %}Homepage{% endblock %}
+        {% block title %}Homepage
+        {% endblock %}
     </title>
+    
 </head>
 <body class="d-flex flex-column min-vh-100">
-
-    <!-- Navigation Bar -->
-    <nav class="nav navbar navbar-expand-lg navbar-dark sticky-top">
+    
+    <!-- including common navbar file on index page  -->
+    <!-- Updated navigation for responsiveness w/ better view -->
+    <nav class="nav navbar navbar-expand-lg navbar-dark sticky-top" >
         <div class="container-fluid">
-            <!-- Logo -->
+            <!-- logo -->
             <img src="{% static 'images/game-logo.png' %}" alt="GamePlan Logo" class="logo">
-            <!-- Toggle button for small screens -->
+            <!-- toggle navigation for the small screen sizes for responsiveness -->
             <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav"
                 aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
                 <span class="navbar-toggler-icon"></span>
             </button>
-    
-            <div class="collapse navbar-collapse" id="navbarNav">
-                <div class="mx-auto">
+
+            <div class="collapse navbar-collapse" id="navbarNav" >
+                <div class="mx-auto"> <!-- This centers the navbar items -->
+                    <!-- navigation list -->
                     <ul class="navbar-nav">
                         <li class="nav-item active"><a class="nav-link" href="{% url 'home' %}">Home</a></li>
                         {% if user.is_authenticated %}
-                        <li class="nav-item dropdown">
-                            <a class="nav-link dropdown-toggle" id="viewToggleDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                                Calendar
-                            </a>
-                            <div class="dropdown-menu" aria-labelledby="viewToggleDropdown">
-                                <a class="dropdown-item" href="{% url 'calendar' user.id %}">Monthly</a>
-                                <a class="dropdown-item" href="{% url 'calendarweek' user.id %}">Weekly</a>
-                            </div>
-                        </li>
-                        <li class="nav-item"><a class="nav-link" href="{% url 'todo_list' %}">To-Do List</a></li>
-                        <li class="nav-item"><a class="nav-link" href="{% url 'user_page' %}">Account</a></li>
-                        <li class="nav-item"><a class="nav-link" href="{% url 'logout' %}">Logout</a></li>
+                        <!-- Calendar-->
+                        <li class="nav-item active"><a class="nav-link" href="{% url 'calendar' user.id %}">Calendar</a></li>
+                            <li class="nav-item"><a class="nav-link" href="{% url 'todo_list' %}">To-Do List</a></li>
+                    
+                            <li class="nav-item"><a class="nav-link" href="{% url 'user_page' %}">Account</a></li>
+                            <li class="nav-item"><a class="nav-link" href="{% url 'logout' %}">Logout</a></li>
                         {% else %}
-                        <li class="nav-item"><a class="nav-link" href="{% url 'login' %}?next={{ request.path }}">Login</a></li>
-                        {% endif %}
+                            <li class="nav-item"><a class="nav-link" href="{% url 'login' %}?next={{ request.path }}">Login</a></li>
+                        {%endif%}
 
                         <!-- Theme Toggle -->
                         <li class="nav-item">
@@ -59,19 +57,22 @@
                                     <span id="theme-icon" class="material-icons" style="cursor: pointer;">light_mode</span>
                                 </label>
                             </div>
-                        </li>                    
+                        </li> 
                     </ul>
-                </div>
+                
+                
+                </div> 
             </div>
         </div>
     </nav>
-
     <!-- Page Content -->
-    <div class="flex-grow-1" id="content"> 
+    <div class="flex-grow-1" id = "content"> 
         {% block content %}{% endblock %}
     </div>
 
-    {% block scripts %}{% endblock %}
+    {% block scripts %}
+    {% endblock %}
+
     {% block footer %}
     <footer class="footer mt-auto">
         <div class="container text-center">
@@ -80,12 +81,15 @@
     </footer>
     {% endblock %}
 
+    
+    
 <!-- Bootstrap and jQuery JS -->
 <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.2/dist/js/bootstrap.bundle.min.js"></script>
 
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.2/dist/js/bootstrap.bundle.min.js"></script>
 <!-- Theme toggle script -->
 <script>
+    
     function toggleTheme() {
         const body = document.body;
         const currentTheme = body.getAttribute('data-theme');
@@ -97,7 +101,7 @@
         // Update the icon
         const themeIcon = document.getElementById('theme-icon');
         themeIcon.textContent = newTheme === 'dark' ? 'light_mode' : 'brightness_2'; // Bright moon for light mode
-
+        
         // Save preference to localStorage
         localStorage.setItem('theme', newTheme);
     }
@@ -111,8 +115,9 @@
         const themeIcon = document.getElementById('theme-icon');
         themeIcon.textContent = savedTheme === 'dark' ? 'light_mode' : 'brightness_2'; // Bright moon for light mode
     });
-</script>
 
+
+</script>
 <style>
     /* Dark and Light Theme Styles */
     body[data-theme='light'] {
@@ -157,7 +162,8 @@
         min-height: 56pxx;
     }
 
-
+    
 </style>
+</div>
 </body>
 </html>
