+++ project/project/settings.py
@@ -37,7 +37,7 @@ ALLOWED_HOSTS = [
     'https://app-hnguye22-5.devedu.io/', 'app-hnguye22-5.devedu.io', 
     'https://app-jfuentes-5.devedu.io/', 'app-jfuentes-5.devedu.io', 
     'https://app-afuina-5.devedu.io/', 'app-afuina-5.devedu.io', 
-    'https://app-smatthe4container-5.devedu.io/', 'app-smatthe4container-5.devedu.io',
+    'https://app-app-workingcontainer-5.devedu.io/', 'app-workingcontainer-5.devedu.io',
 ]
 
 CSRF_TRUSTED_ORIGINS = [
@@ -181,4 +181,3 @@ LOGOUT_REDIRECT_URL = '/'  # Adjusts prebuilt logout path to redirect user to in
 #https://www.geeksforgeeks.org/python-uploading-images-in-django/
 MEDIA_ROOT =  os.path.join(BASE_DIR, 'media')
 MEDIA_URL = '/media/'
-
